Issue/234: Renamed "drivers/fate" tab to "personality" tab
* Also renamed and moved templates around as appropriate.
