+++ presentation/lang/en.json
@@ -498,7 +498,7 @@
         "tab": {
           "abilities": "Abilities",
           "magic": "Magic",
-          "driversFate": "Drivers/Fate",
+          "personality": "Personality",
           "health": "Health",
           "assets": "Assets",
           "biography": "Biography",
