+++ presentation/sheet/actor/actor-sheet-viewmodel.mjs
@@ -6,7 +6,7 @@ import ViewModel from "../../view-model/view-model.mjs"
 import ActorAssetsViewModel from "./part/assets/actor-assets-viewmodel.mjs"
 import ActorAttributesViewModel from "./part/abilities/actor-attributes-viewmodel.mjs"
 import ActorSkillsViewModel from "./part/abilities/actor-skills-viewmodel.mjs"
-import ActorDriversFateViewModel from "./part/actor-drivers-fate-viewmodel.mjs"
+import ActorPersonalityViewModel from "./part/personality/actor-personality-viewmodel.mjs"
 import ActorBiographyViewModel from "./part/actor-biography-viewmodel.mjs"
 import ActorHealthViewModel from "./part/health/actor-health-viewmodel.mjs"
 import ActorPersonalsViewModel from "./part/actor-personals-viewmodel.mjs"
@@ -18,7 +18,7 @@ import BaseSheetViewModel from "../../view-model/base-sheet-viewmodel.mjs"
  * @property {ViewModel} personalsViewModel
  * @property {LazyLoadViewModel} attributesViewModel
  * @property {LazyLoadViewModel} skillsViewModel
- * @property {LazyLoadViewModel} driversFateViewModel
+ * @property {LazyLoadViewModel} personalityViewModel
  * @property {LazyLoadViewModel} healthViewModel
  * @property {LazyLoadViewModel} assetsViewModel
  * @property {LazyLoadViewModel} biographyViewModel
@@ -137,11 +137,11 @@ export default class ActorSheetViewModel extends BaseSheetViewModel {
         },
       });
       if (args.document.type === 'pc') {
-        this.driversFateViewModel = new LazyLoadViewModel({
-          id: "lazyDriversFate",
+        this.personalityViewModel = new LazyLoadViewModel({
+          id: "lazyPersonality",
           parent: thiz,
-          template: TEMPLATES.ACTOR_DRIVERS_FATE,
-          viewModelFactoryFunction: (args) => { return new ActorDriversFateViewModel(args); },
+          template: TEMPLATES.ACTOR_PERSONALITY,
+          viewModelFactoryFunction: (args) => { return new ActorPersonalityViewModel(args); },
           viewModelArgs: {
             ...args, 
             id: "drivers-fate", 
@@ -257,7 +257,7 @@ export default class ActorSheetViewModel extends BaseSheetViewModel {
       await this.attributesViewModel.render();
       await this.skillsViewModel.render();
     } else if (tab === "drivers-fate") {
-      await this.driversFateViewModel.render();
+      await this.personalityViewModel.render();
     } else if (tab === "health") {
       await this.healthViewModel.render();
     } else if (tab === "assets") {
