+++ presentation/sheet/actor/actor-sheet.hbs
@@ -30,7 +30,7 @@ cssClass: {undefined | String}
         {{!-- Default tab is specified in actor-sheet.mjs --}}
         <a class="item flex flex-center flex-grow{{#if viewModel.showFancyFont}} fancy-font{{/if}}" data-tab="abilities"><div>{{localize "ambersteel.character.sheet.tab.abilities"}}</div></a>
         {{#if viewModel.isPC }}
-          <a class="item flex flex-center flex-grow{{#if viewModel.showFancyFont}} fancy-font{{/if}}" data-tab="drivers-fate"><div>{{localize "ambersteel.character.sheet.tab.driversFate"}}</div></a>
+          <a class="item flex flex-center flex-grow{{#if viewModel.showFancyFont}} fancy-font{{/if}}" data-tab="drivers-fate"><div>{{localize "ambersteel.character.sheet.tab.personality"}}</div></a>
         {{/if}}
         <a class="item flex flex-center flex-grow{{#if viewModel.showFancyFont}} fancy-font{{/if}}" data-tab="health"><div>{{localize "ambersteel.character.sheet.tab.health"}}</div></a>
         <a class="item flex flex-center flex-grow{{#if viewModel.showFancyFont}} fancy-font{{/if}}" data-tab="assets"><div>{{localize "ambersteel.character.sheet.tab.assets"}}</div></a>
@@ -61,7 +61,7 @@ cssClass: {undefined | String}
       {{#if viewModel.isPC }}
       {{!-- Driverss & Fate Tab --}}
         <div class="tab drivers-fate" data-group="primary" data-tab="drivers-fate">
-          {{> lazy viewModel=viewModel.driversFateViewModel}}
+          {{> lazy viewModel=viewModel.personalityViewModel}}
         </div>
       {{/if}}
       {{!-- Health Tab --}}
