+++ presentation/sheet/actor/part/personality/actor-personality.hbs
@@ -1,5 +1,5 @@
 {{!-- Parameters:
-viewModel: {ActorDriversFateViewModel}
+viewModel: {ActorPersonalityViewModel}
 cssClass: {undefined | String}
 --}}
 <section class="{{cssClass}}">
