+++ presentation/templatePreloader.mjs
@@ -87,9 +87,9 @@ export const TEMPLATES = {
   ACTOR_ATTRIBUTE_TABLE: `${basePathPresentation}/sheet/actor/part/abilities/actor-attribute-table.hbs`,
   ACTOR_ATTRIBUTES: `${basePathPresentation}/sheet/actor/part/abilities/actor-attributes.hbs`,
   ACTOR_SKILLS: `${basePathPresentation}/sheet/actor/part/abilities/actor-skills.hbs`,
-  ACTOR_DRIVERS: `${basePathPresentation}/sheet/actor/part/actor-drivers.hbs`,
+  ACTOR_DRIVERS: `${basePathPresentation}/sheet/actor/part/personality/actor-drivers.hbs`,
   ACTOR_FATE: `${basePathPresentation}/sheet/actor/part/actor-fate.hbs`,
-  ACTOR_DRIVERS_FATE: `${basePathPresentation}/sheet/actor/part/actor-drivers-fate.hbs`,
+  ACTOR_PERSONALITY: `${basePathPresentation}/sheet/actor/part/personality/actor-personality.hbs`,
   ACTOR_HEALTH: `${basePathPresentation}/sheet/actor/part/health/actor-health.hbs`,
   ACTOR_HEALTH_STATES: `${basePathPresentation}/sheet/actor/part/health/actor-health-states.hbs`,
   ACTOR_HEALTH_STATES_LIST_ITEM: `${basePathPresentation}/sheet/actor/part/health/actor-health-states-list-item.hbs`,
